adding kotlin model classes
