+++ core/baker-compiler/src/main/kotlin/com/ing/baker/compiler/PreCompileValidations.kt
@@ -1,10 +1,10 @@
 package com.ing.baker.compiler
 
-import com.ing.baker.model.Event
-import com.ing.baker.model.Ingredient
-import com.ing.baker.model.Interaction
-import com.ing.baker.model.Recipe
-import com.ing.baker.model.toKotlin
+import com.ing.baker.recipe.Event
+import com.ing.baker.recipe.Ingredient
+import com.ing.baker.recipe.Interaction
+import com.ing.baker.recipe.Recipe
+import com.ing.baker.recipe.toKotlin
 import scala.jdk.javaapi.CollectionConverters
 import com.ing.baker.recipe.common.Recipe as ScalaRecipe
 
@@ -46,7 +46,7 @@ object PreCompileValidations {
 
     private fun assertSensoryEventsNegativeFiringLimits(recipe: Recipe): List<String> =
         recipe.sensoryEvents
-            .firstOrNull { it.maxFiringLimit != null && it.maxFiringLimit <= 0 }
+            .firstOrNull { it.maxFiringLimit != null && it.maxFiringLimit!! <= 0 }
             ?.let { listOf("MaxFiringLimit should be greater than 0") } ?: emptyList()
 
     private fun assertRequiredEventForReprovider(recipe: Recipe): List<String> =
