+++ core/baker-compiler/src/main/kotlin/com/ing/baker/recipe/CheckPointEvent.kt
@@ -1,0 +1,8 @@
+package com.ing.baker.recipe
+
+data class CheckPointEvent(
+    val name: String = "",
+    val requiredEvents: Set<String> = emptySet(),
+    val requiredOneOfEvents: Set<Set<String>> = emptySet(),
+)
+
