+++ core/baker-compiler/src/main/kotlin/com/ing/baker/recipe/Event.kt
@@ -1,0 +1,7 @@
+package com.ing.baker.recipe
+
+data class Event(
+    val name: String,
+    val providedIngredients: List<Ingredient>,
+    val maxFiringLimit: Int? = null,
+)
