+++ core/baker-compiler/src/main/kotlin/com/ing/baker/recipe/EventOutputTransformer.kt
@@ -1,0 +1,3 @@
+package com.ing.baker.recipe
+
+data class EventOutputTransformer(val newEventName: String, val ingredientRenames: Map<String, String>)
