+++ core/baker-compiler/src/main/kotlin/com/ing/baker/recipe/Ingredient.kt
@@ -1,0 +1,5 @@
+package com.ing.baker.recipe
+
+import com.ing.baker.types.Type
+
+data class Ingredient(val name: String, val type: Type)
