+++ core/baker-compiler/src/main/kotlin/com/ing/baker/recipe/Interaction.kt
@@ -1,49 +1,8 @@
-package com.ing.baker.model
+package com.ing.baker.recipe
 
 import com.ing.baker.recipe.common.InteractionFailureStrategy
-import com.ing.baker.types.Type
-import kotlin.collections.mapValues
-import kotlin.collections.toMap
-import kotlin.time.Duration
-import java.lang.reflect.Type as JavaType
 import com.ing.baker.types.Converters
 import com.ing.baker.types.Value
-
-data class Event(
-    val name: String,
-    val providedIngredients: List<Ingredient>,
-    val maxFiringLimit: Int? = null,
-)
-
-data class CheckPointEvent(
-    val name: String = "",
-    val requiredEvents: Set<String> = emptySet(),
-    val requiredOneOfEvents: Set<Set<String>> = emptySet(),
-)
-
-data class Ingredient(val name: String, val type: Type)
-
-data class Recipe(
-    val name: String,
-    val interactions: List<Interaction>,
-    val sensoryEvents: Set<Event>,
-    val subRecipes: Set<Recipe>,
-    val defaultFailureStrategy: InteractionFailureStrategy,
-    val eventReceivePeriod: Duration? = null,
-    val retentionPeriod: Duration? = null,
-    val checkpointEvents: Set<CheckPointEvent>,
-    val sieves: Set<Sieve>,
-)
-
-data class Sieve(
-    val name: String = "",
-    val inputIngredients: List<Ingredient>,
-    val output: List<Event>,
-    val function: Any,
-    val javaTypes: List<JavaType>
-)
-
-data class EventOutputTransformer(val newEventName: String, val ingredientRenames: Map<String, String>)
 
 data class Interaction(
     val name: String,
@@ -64,30 +23,23 @@ data class Interaction(
         @JvmStatic
         fun of(
             name: String,
-            interaction: Interaction,
-            requiredEvents: Set<String>,
-            requiredOneOfEvents: Set<Set<String>>,
-            predefinedIngredients: Map<String, Any>,
-            overriddenIngredientNames: Map<String, String>,
-            eventOutputTransformers: Map<Event, EventOutputTransformer>,
-            maximumInteractionCount: Int?,
-            failureStrategy: InteractionFailureStrategy?,
-            isReprovider: Boolean,
+            inputIngredients: List<Ingredient>,
+            output: List<Event>
         ): Interaction =
             Interaction(
                 name,
-                interaction.originalName,
-                interaction.inputIngredients,
-                interaction.output,
-                requiredEvents,
-                requiredOneOfEvents,
-                predefinedIngredients.mapValues { Converters.toValue(it.value) },
-                overriddenIngredientNames,
+                originalName = name,
+                inputIngredients,
+                output,
+                requiredEvents = emptySet(),
+                requiredOneOfEvents = emptySet(),
+                predefinedIngredients = emptyMap(),
+                overriddenIngredientNames = emptyMap(),
                 overriddenOutputIngredientName = null,
-                eventOutputTransformers,
-                maximumInteractionCount,
-                failureStrategy,
-                isReprovider
+                eventOutputTransformers = emptyMap(),
+                maximumInteractionCount = null,
+                failureStrategy = null,
+                isReprovider = false
             )
 
         @JvmStatic
